+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to this project will be documented in this file.
 ### New
 
 - support cookies in net module for std mode (not wasm)
-
+- remove network aliases (main, dev, main.ton.dev, net.ton.dev)
 
 ## [1.37.2] – 2022-08-10