+++ CHANGELOG.md
@@ -6,7 +6,12 @@ All notable changes to this project will be documented in this file.
 
 ### New
 
-- **Debot module**:
+- Supported cookies in net module for std mode (not wasm)
+- Removed network aliases (main, dev, main.ton.dev, net.ton.dev)
+- Removed separate rempEnabled query.
+- Changed 401 error message to response text (was "Unauthorized").
+
+**Debot module**:
     - ABI specification v2.3 is supported in DEngine.
     - Supported flags `OVERRIDE_TS`, `OVERRIDE_EXPT`, `ASYNC_CALL` for external messages in DEngine.