+++ CHANGELOG.md
@@ -14,9 +14,10 @@ All notable changes to this project will be documented in this file.
 
 - Support cookies in net module for std mode (not wasm)
 - Remove network aliases (main, dev, main.ton.dev, net.ton.dev)
-- Removed the check of REMP support on backend during client initialization. 
-  It will make client initialization faster -> CLI tools that use SDK will work faster, web pages will load initial data faster.
+- No balancing logic in case of 1 endpoint + removed the check of REMP support on backend during client initialization. 
+  These changes will make client initialization faster -> CLI tools that use SDK will work faster, web pages will load initial data faster.
 - Changed 401 error message to response message from API
+- Tests improvements: cryptobox tests made stable
 
 ## [1.37.2] – 2022-08-10