+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [1.38.1] – 2022-11-06
+## [1.38.1] – 2022-11-10
 
 ### Improvement